Set up containers

+++ start.sh
@@ -6,18 +6,24 @@ PW="s3cr3t";
 DB="lucky_challenge";
 REDIS_SERVER="lucky_redis"
 API_SERVER="lucky_app"
+MIGRATIONS_SERVER="lucky_migrations"
+RUN_MIGRATIONS="true"
+CLEAN_CONTAINERS="true"
 
-
-echo "Removing Previous Containers"
-(docker kill $SERVER || :) && \
-  (docker rm $SERVER || :)
-(docker kill $SERVER || :) && \
-  (docker rm $SERVER || :)
-(docker kill $REDIS_SERVER || :) && \
-  (docker rm $REDIS_SERVER || :)
-(docker kill $API_SERVER || :) && \
-  (docker rm $API_SERVER || :)
-docker network remove lucky_challenge_net
+if [ "$CLEAN_CONTAINERS" == "true" ]; then
+  echo "Removing Previous Containers"
+  (docker kill $SERVER || :) && \
+    (docker rm $SERVER || :)
+  (docker kill $SERVER || :) && \
+    (docker rm $SERVER || :)
+  (docker kill $REDIS_SERVER || :) && \
+    (docker rm $REDIS_SERVER || :)
+  (docker kill $MIGRATIONS_SERVER || :) && \
+    (docker rm $MIGRATIONS_SERVER || :)
+  (docker kill $API_SERVER || :) && \
+    (docker rm $API_SERVER || :)
+  docker network remove lucky_challenge_net
+fi
 
 echo "Creating network"
 docker network create lucky_challenge_net
@@ -43,9 +49,11 @@ docker run --net lucky_challenge_net --name $REDIS_SERVER \
   -p 6379:6379 \
   -d redis
 
-# wait for pg to start
-echo "sleep wait for redis [$REDIS_SERVER] to start";
-SLEEP 3;
+if [ "$RUN_MIGRATIONS" == "true" ]; then
+  echo "RUNNING MIGRATIONS";
+  docker build -t $MIGRATIONS_SERVER -f ./migrations.Dockerfile .
+  docker run --net lucky_challenge_net --name $MIGRATIONS_SERVER -d $MIGRATIONS_SERVER
+fi
 
 echo "SETTING UP API"
 docker build -t $API_SERVER .
